requirements (1).txt
easydict==1.11
gradio==4.1.1
gradio-client==0.7.0
huggingface-hub==0.18.0
matplotlib-inline==0.1.6
mediapipe==0.10.7
numpy==1.24.4
opencv-contrib-python==4.8.1.78
pandas==2.0.3
pillow==10.1.0
scikit-learn==1.3.2
seaborn==0.13.0
torchvision==0.16.0
PyWavelets

